requirements.txt
streamlit
yt-dlp
youtube-transcript-api
moviepy
google-generativeai
python-dotenv
ffmpeg-python
